Organized folder structure for all 3 tests
