+++ bundle-workflow/tests/test_perf_workflow/test_perf_test_cluster.py
@@ -9,7 +9,7 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from manifests.bundle_manifest import BundleManifest
-from test_workflow.perf_test_cluster import PerfTestCluster
+from test_workflow.perf_test.perf_test_cluster import PerfTestCluster
 
 
 class TestPerfTestCluster(unittest.TestCase):
